# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lobsense IoT V2 API Documentation",
        "description": "Dokumentasi resmi API sistem monitoring, telemetri sensor tambak, pengontrolan aerator/feeder, inferensi AI, dan manajemen perangkat Lobsense V2.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lobsense API Local Server"
        }
    ],
    "paths": {
...
        "/api/v2/system-settings": {
            "get": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Pengaturan Sistem",
                "description": "Mengambil seluruh pengaturan sistem termasuk lokasi, logo, dan konfigurasi.",
                "operationId": "ba81ffb5fe7b23d90ab5b8e70083cfe9",
                "responses": {
                    "200": {
                        "description": "Pengaturan sistem berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "System Settings"
                ],
                "summary": "Update Pengaturan Sistem",
                "description": "Memperbarui pengaturan sistem termasuk lokasi, nama instansi, dan upload logo.",
                "operationId": "6c004d7da411bbbdcd444c72de24d20f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "system_latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": -8.583
                                    },
                                    "system_longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 116.116
                                    },
                                    "system_city_name": {
                                        "type": "string",
                                        "example": "Lombok Barat"
                                    },
                                    "system_province_code": {
                                        "type": "string",
                                        "example": "52"
                                    },
                                    "system_city_code": {
                                        "type": "string",
                                        "example": "5201"
                                    },
                                    "system_district_code": {
                                        "type": "string",
                                        "example": "520101"
                                    },
                                    "system_logo_text": {
                                        "type": "string",
                                        "example": "Lobsense"
                                    },
                                    "system_instansi_name": {
                                        "type": "string",
                                        "example": "PT Lobster Sensing Indonesia"
                                    },
                                    "system_logo_file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "logo_file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengaturan berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/thresholds": {
            "get": {
                "tags": [
                    "Thresholds"
                ],
                "summary": "Daftar Ambang Batas Sensor",
                "description": "Mengambil konfigurasi ambang batas sensor untuk IoT Node tertentu berdasarkan serial number.",
                "operationId": "136dc1a702dd5e7c6474a7c3951370b2",
                "parameters": [
                    {
                        "name": "iot_node_serial_number",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "DEMO-NODE-001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar ambang batas berhasil diambil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/thresholds/bulk-update": {
            "post": {
                "tags": [
                    "Thresholds"
                ],
                "summary": "Bulk Update Ambang Batas Sensor",
                "description": "Memperbarui atau membuat ambang batas sensor secara massal untuk satu IoT Node.",
                "operationId": "e253d697fa462a52a66253d0b71d83fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "iot_node_serial_number",
                                    "thresholds"
                                ],
                                "properties": {
                                    "iot_node_serial_number": {
                                        "type": "string",
                                        "example": "DEMO-NODE-001"
                                    },
                                    "thresholds": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "sensor_code": {
                                                    "type": "string",
                                                    "example": "ph"
                                                },
                                                "value_min": {
                                                    "type": "number",
                                                    "example": 6.5
                                                },
                                                "value_max": {
                                                    "type": "number",
                                                    "example": 8.5
                                                },
                                                "offset_value": {
                                                    "type": "number",
                                                    "example": 0.1
                                                },
                                                "filter_rules": {
                                                    "type": "string",
                                                    "example": "moving_average"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ambang batas berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/weather/latest": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Data Cuaca Terbaru",
                "description": "Mengambil laporan cuaca terbaru dengan filter opsional berdasarkan kode kota.",
                "operationId": "e6bcce13714499ef11e4e5e41643c581",
                "parameters": [
                    {
                        "name": "city_code",
                        "in": "query",
                        "description": "Kode kota untuk filter cuaca",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "5201"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data cuaca terbaru berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}